refactor: 견적 확정 + 견적 반려 호출 방식 변경 및 타입 수정
