+++ src/types/lesson.ts
@@ -1,4 +1,4 @@
-import { Profile } from "@/types/types";
+import { LessonRequestStatus, LessonType, LocationType, Profile } from "@/types/types";
 
 export interface LessonParams {
   page?: number;
@@ -68,18 +68,18 @@ export interface LessonResult {
 export interface MyLesson {
   id: string;
   userId: string;
-  lessonType: string;
+  lessonType: LessonType;
   lessonSubType?: string;
   startDate: string;
   endDate: string;
   lessonCount: number;
   lessonTime: number;
   quoteEndDate: string;
-  locationType: string;
+  locationType: LocationType;
   postcode?: string;
   roadAddress?: string;
   detailAddress?: string;
-  status: string;
+  status: LessonRequestStatus;
   createdAt: string;
   updatedAt: string;
   directQuoteRequests: [
